Split up wagtail.admin.views.generic module
