+++ wagtail/admin/views/generic/__init__.py
@@ -1,0 +1,3 @@
+from .hooks import HookResponseMixin  # noqa
+from .models import CreateView, DeleteView, EditView, IndexView  # noqa
+from .permissions import PermissionCheckedMixin  # noqa
